+++ src/middlewares/auth.js
@@ -1,17 +1,12 @@
 export const auth = function (req, res, next) {
-    const { user, password } = req.body;
-    
-    if (user === "admincoder@coder.com" && password === "adminCod3r123") {
-        req.session.user = user;
-        req.session.admin = true;
-        const adminUser = {
-            user: "admincoder@coder.com",
-            role: "admin"
-        }
-
-        req.session.user = adminUser;
-        return res.redirect("/");
+    const { user } = req.body;
+  
+    if (req.user && req.user.role === 'admin' && req.path !== '/') {
+      req.session.user = user;
+      req.session.admin = true;
+  
+      return res.redirect("/");
     } else {
-        return next();
+      return next();
     }
-}
+  }