+++ src/routes/productsRoute.js
@@ -2,6 +2,7 @@ import { Router } from "express";
 import { uploader } from "../utils/multerUtil.js";
 import ProductController from "../controllers/productController.js";
 import { productService } from "../repositories/index.js";
+import { isAdmin } from "../middlewares/guard.js";
 
 const productsRouter = Router();
 const myProduct = new ProductController(productService);
@@ -23,7 +24,7 @@ productsRouter.get('/:pid', async (req, res) => {
     }
 });
 
-productsRouter.post("/", uploader.array('thumbnail', 3), async (req, res) => {
+productsRouter.post("/", uploader.array('thumbnail', 3), isAdmin, async (req, res) => {
     if (req.files) {
         const thumbnails = req.files.map((file) => file.filename);
         req.body.thumbnail = thumbnails;
@@ -37,7 +38,7 @@ productsRouter.post("/", uploader.array('thumbnail', 3), async (req, res) => {
     }
 });
 
-productsRouter.put("/:pid", uploader.array('thumbnail', 3), async (req, res) => {
+productsRouter.put("/:pid", uploader.array('thumbnail', 3), isAdmin, async (req, res) => {
     try {
         const pid = req.params.pid;
     
@@ -66,7 +67,7 @@ productsRouter.put("/:pid", uploader.array('thumbnail', 3), async (req, res) => 
     }
   });
 
- productsRouter.delete("/:pid", async (req, res) => {
+ productsRouter.delete("/:pid", isAdmin, async (req, res) => {
     try {
         const pid = req.params.pid;
         const product = await myProduct.getProductById(pid);